* tp/Texinfo/XS/main/convert_to_text.h (TEXT_OPTIONS): use size_t for document_descriptor.
* tp/Texinfo/XS/parsetexi/parser.c (parse_texi_document): return void.

* tp/Texinfo/XS/convert/convert_html.c,
tp/Texinfo/XS/main/DocumentXS.xs,
tp/Texinfo/XS/main/build_perl_info.c, tp/Texinfo/XS/main/document.c,
tp/Texinfo/XS/main/get_perl_info.c, tp/Texinfo/XS/main/translations.c,
tp/Texinfo/XS/parsetexi/Parsetexi.xs, tp/Texinfo/XS/parsetexi/api.c,
tp/Texinfo/XS/parsetexi/parser.c: use size_t more for document
descriptors.

+++ tp/Texinfo/XS/main/build_perl_info.c
@@ -17,6 +17,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
+#include <stddef.h>
 
 /* Avoid namespace conflicts. */
 #define context perl_context
@@ -1577,7 +1578,7 @@ pass_errors_to_registrar (const ERROR_MESSAGE_LIST *error_messages,
 }
 
 void
-pass_document_parser_errors_to_registrar (int document_descriptor,
+pass_document_parser_errors_to_registrar (size_t document_descriptor,
                                           SV *parser_sv)
 {
   DOCUMENT *document;
